Reject on invalid response body.
Certain situations currently cause the Postageapp API to respond with an HTML error page rather than a JSON response body.
This caused the Postageapp send() method to throw a SyntaxError attempting to parse the response body rather than rejecting
the Promise.

If merged, this PR will catch the error thrown by JSON.parse() and reject with the caught error.